Cart function
Continued on the cart function and view cart.
Now the site renders items pushed into an array from localStorage and displays them on the site for the user. User is also able to remove items from the cart.

Will have to update price for products on next build as some items is on sale.
Also have to add a total price section and buy button so the user can complete the order.